scp.c: cast strlen() calls
```
src/scp.c:301:..: warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
src/scp.c:301:19: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
src/scp.c:301:19: warning: conversion from 'size_t' {aka 'long unsigned int'} to 'int' may change value [-Wconversion]
src/scp.c:301:19: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
src/scp.c:301:19: warning: implicit conversion loses integer precision: 'unsigned long' to 'int' [-Wshorten-64-to-32]
src/scp.c:860:..: warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
src/scp.c:860:19: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'int' may change value [-Wconversion]
src/scp.c:860:19: warning: conversion from 'size_t' {aka 'long unsigned int'} to 'int' may change value [-Wconversion]
src/scp.c:860:19: warning: implicit conversion loses integer precision: 'unsigned long long' to 'int' [-Wshorten-64-to-32]
src/scp.c:860:19: warning: implicit conversion loses integer precision: 'unsigned long' to 'int' [-Wshorten-64-to-32]
```

+++ src/scp.c
@@ -298,7 +298,7 @@ scp_recv(LIBSSH2_SESSION * session, const char *path, libssh2_struct_stat * sb)
                  session->scpRecv_command_len,
                  "scp -%sf ", sb?"p":"");
 
-        cmd_len = strlen((char *)session->scpRecv_command);
+        cmd_len = (int)strlen((char *)session->scpRecv_command);
         cmd_len += shell_quotearg(path,
                                   &session->scpRecv_command[cmd_len],
                                   session->scpRecv_command_len - cmd_len);
@@ -859,7 +859,7 @@ scp_send(LIBSSH2_SESSION * session, const char *path, int mode,
                  session->scpSend_command_len,
                  "scp -%st ", (mtime || atime)?"p":"");
 
-        cmd_len = strlen((char *)session->scpSend_command);
+        cmd_len = (int)strlen((char *)session->scpSend_command);
         cmd_len += shell_quotearg(path,
                                   &session->scpSend_command[cmd_len],
                                   session->scpSend_command_len - cmd_len);
